+++ plugins/migration/copytable/copytable.cpp
@@ -758,7 +758,7 @@ SQLRETURN ODBCCopyDataSource::get_geometry_buffer_data(RowBuffer &rowbuffer, int
 
       if (inbuf_len > 0)
         log_warning("%lu characters could not be converted to UTF-8 from column %s during copy\n",
-                    inbuf_len, (*_columns)[column-1].source_name.c_str());
+                    (unsigned long)inbuf_len, (*_columns)[column-1].source_name.c_str());
 
       *out_length = out_buffer_len - outbuf_len;
     }
@@ -2364,7 +2364,7 @@ void MySQLCopyDataTarget::backup_triggers_for_schema(const std::string &schema)
       }
     }
 
-    log_info("Successfully backed up %lu triggers.\n", triggers.size());
+    log_info("Successfully backed up %lu triggers.\n", (unsigned long)triggers.size());
   }
 }
 
@@ -2456,7 +2456,7 @@ void MySQLCopyDataTarget::restore_triggers(std::set<std::string> &schemas)
       // Drops the trigger backup table
       drop_trigger_backups(a_schema);
 
-      log_info("Trigger Restore: %d succeeded, %lu failed\n", restored, trigger_sql.size() - restored);
+      log_info("Trigger Restore: %d succeeded, %lu failed\n", restored, (unsigned long)(trigger_sql.size() - restored));
     }
   }
 }
